IGAPP-136: Move InternalPathnameParser to api client
